CHANGE data types from Double to Int, ADD level of difficulty to recipes

+++ Client/Client/Models/Ingredient.swift
@@ -12,7 +12,7 @@ struct Ingredient: Codable, Hashable {
     /// the ingredient's name
     let name: String
     /// The amount or quantity of the ingredient that is required for the recipe
-    let quantity: Double
+    let quantity: Int
     /// The unit used to measure the amount of the ingredient needed for the recipe
     let unit: String
 }
